Fixing an issue with deleting nested S3 folders
When deleting a nested folder, we first list the objects in the folder
using a key prefix. The key prefix was actually returning the place
holder key created in the AWS Management Console (e.g., 'foo/'), which
failed the test that ensures the directory is empty. We are now delting
the placeholder key and ensuring it is the only key in the folder.

Closes #328.