Replacing phing with simple PHP script.
This removes a dependency on phing, which was only used to build the
phar and zip. This change adds a simple PHP script that is used to
create the same exact zip and phar files, but does so without a
dependency on phing and only has to build the staging directory once.

This change is a step along the way of automating Github release
deployments via travis-ci as it solves the problem of how to give travis
access to phing.