Adding a system for more helpful factory errors.
This commit adds a way for us to create custom error messages for
missing required client configuration options. In this commit, I've
added helpful error messages for a missing region (i.e., a link to
rande.html), and a helpful error message for a missing version (i.e., a
list of available versions on disk and a link to the API docs).