Making providers more consistent and composable.
This commit updates the CredentialsProvider, SignatureProvider, and
EndpointProvider to be more consistent. The providers now are all
functions that accept arguments and return a value or null. You can use
Aws\Utils::orFn() to compose multiple providers into a single provider.
Each provider class has static functions used to create providers and
contains a `resolve()` function that invokes a provider and ensures it
returns the appropriate type of value or throws a specific exception.