Modifying Sdk class to use namespace name.
The Aws\Sdk class will now use namespace names to resolve clients. This
removes the need for endpoint prefix mapping, and means we only need to
provide a hash of aliases to endpoint prefixes where they differ (e.g.,
"email"). I've also update the getClient() and get*() methods to use the
word create to better communicate that the result is not cached but
rather creates a new client each time it is invoked.