Updating validation to be more flexible
This commit updates the SDK to allow customers to configure exactly what
they want to validate client side. I've added the ability to provide
validation constraints as feature flags to a Validator and to the
validate client option. This commit also updates the SDK to no longer
validate "max" by default.