Fix DynamoDB session handler sleep time (#1122)
Between attempts to access a locked session, the handler is meant to
wait for a configurable number of microseconds. The handler code was
previously multiplying the wait time by 10^-6 before passing it to
usleep, even though the values were already in microseconds. For the
default min and max wait time values of 10000 and 50000, the usleep
argument would always be less than one and so the handler wouldn't
wait at all between attempts.

This commit removes the multiplication by 10^-6, so the handler sleeps
for the intended amount of time between requests.