Fixes validation on assoc arrays having a 0 index (#1732)
* Fixes validation on assoc arrays having a 0 index

The existing heuristic for determining an assoc array looks only
for the presence of a `0` index.  However to correctly determine
whether a php array will be serialized (json) or marshaled (DynamoDB)
into a list (array) or map (assoc), the hueristic must be to check
for sequential integer keys starting at 0.

This change will walk the array checking the key against an incrementing
integer.  at any point where the key does not meet the expected integer
value, the iteration will stop and the array will be considered
associative. (this implementation is similar to the DynamoDB marshaler)

For any array not having a 0 index, this method will return after
the first iteration, so for the shape of the assoc arrays that this
method successfully validated before this change, there will be little
to no performance impact.

the only difference will be for arrays that have a 0 index, which this
method did NOT properly validate previously so there is no baseline
for a performance comparison.

https://github.com/aws/aws-sdk-php/issues/1722

* Clarify variable names and tests

Response to PR feedback

Resolves #1722